+++ crack-core/src/commands/admin/mod.rs
@@ -1,0 +1,39 @@
+pub mod admin;
+pub mod authorize;
+pub mod ban;
+pub mod broadcast_voice;
+pub mod create_role;
+pub mod create_text_channel;
+pub mod create_voice_channel;
+pub mod deauthorize;
+pub mod delete_channel;
+pub mod kick;
+pub mod mute;
+pub mod print_settings;
+pub mod set_all_log_channel;
+pub mod set_all_log_channel_data;
+pub mod set_all_log_channel_old_data;
+pub mod set_prefix;
+pub mod set_welcome_settings;
+pub mod unban;
+pub mod unmute;
+
+pub use admin::*;
+pub use authorize::*;
+pub use ban::*;
+pub use broadcast_voice::*;
+pub use create_role::*;
+pub use create_text_channel::*;
+pub use create_voice_channel::*;
+pub use deauthorize::*;
+pub use delete_channel::*;
+pub use kick::*;
+pub use mute::*;
+pub use print_settings::*;
+pub use set_all_log_channel::*;
+pub use set_all_log_channel_data::*;
+pub use set_all_log_channel_old_data::*;
+pub use set_prefix::*;
+pub use set_welcome_settings::*;
+pub use unban::*;
+pub use unmute::*;
